refactor(core): extract createDragSelectors
Pull the 10 viewport-aware drag/shift-multi-select derivations out of
init() into a focused selectors module:

  - pendingShiftSelectionElement / pendingShiftSelectionBounds
  - isDraggingBeyondThreshold
  - dragBounds (single drag-rect bounds)
  - dragPreviewBounds (per-element preview during drag)
  - selectionBoundsMultiple (preview > pending+frozen > frozen)
  - frozenLabelEntries (per-frozen-element labels)
  - pendingShiftPreviewEntry (about-to-add label)
  - cursorPosition (pointer or copy-anchored)
  - shiftSelectionLabelMouseX

These all share the viewport-version-aware pattern and the
element-connectedness guard, plus depend on the same set of inputs
(grab, phase, elementSelectors, shiftMultiSelect controller,
isPendingContextMenuSelect, debouncedDragPointer). The space-drag
repositioning helpers (isSpaceActivationKey, startSpaceDragRepositioning,
stopSpaceDragRepositioning) stay in init() — they couple to the
previousSpaceDragPointerPage closure flag that the drag handlers
still own.

Drops 4 now-unused imports (mapArray, isRootElement, createFlatOverlayBounds,
FrozenLabelEntry type, viewportVersion destructure).

index.tsx: 2074 -> 1952 lines (-122). For the first time the file is
under 2000 lines. Cumulative since the start of the PR: 3724 -> 1952
(-1772 / -47.6%). 54 shift-multi-select + drag-selection + visual-
feedback specs green.
