+++ packages/react-grab/src/core/index.tsx
@@ -8,7 +8,6 @@ import {
   createEffect,
   createResource,
   on,
-  mapArray,
 } from "solid-js";
 import { render } from "solid-js/web";
 import { createGrabStore } from "./store.js";
@@ -35,6 +34,7 @@ import { createPromptModeHandlers } from "./prompt-mode-handlers.js";
 import { createWindowFocusListeners } from "./window-focus-listeners.js";
 import { createToolbarStateController } from "./toolbar-state-controller.js";
 import { createShiftMultiSelectState } from "./shift-multi-select-state.js";
+import { createDragSelectors } from "./drag-selectors.js";
 import {
   isKeyboardEventTriggeredByInput,
   hasTextSelectionInInput,
@@ -55,13 +55,11 @@ import {
   getElementsAtPoint,
 } from "../utils/get-element-at-position.js";
 import { isValidGrabbableElement } from "../utils/is-valid-grabbable-element.js";
-import { isRootElement } from "../utils/is-root-element.js";
 import { isElementConnected } from "../utils/is-element-connected.js";
 import { getElementsInDrag } from "../utils/get-elements-in-drag.js";
 import { getElementAnchorRatio } from "../utils/get-element-anchor-ratio.js";
 import { createElementBounds } from "../utils/create-element-bounds.js";
 import {
-  createFlatOverlayBounds,
   createPageRectFromBounds,
 } from "../utils/create-bounds-from-drag-rect.js";
 import { getTagName } from "../utils/get-tag-name.js";
@@ -95,7 +93,6 @@ import type {
   ReactGrabState,
   ContextMenuActionContext,
   ContextMenuAction,
-  FrozenLabelEntry,
   SettableOptions,
   SourceInfo,
   Plugin,
@@ -161,7 +158,7 @@ export const init = (rawOptions?: Options): ReactGrabAPI => {
       theme: DEFAULT_THEME,
       keyHoldDuration: pluginRegistry.store.options.keyHoldDuration ?? DEFAULT_KEY_HOLD_DURATION_MS,
     });
-    const { store, actions, pointer, viewportVersion } = grab;
+    const { store, actions, pointer } = grab;
 
     const phase = createGrabPhaseSelectors(grab);
     const {
@@ -331,36 +328,30 @@ export const init = (rawOptions?: Options): ReactGrabAPI => {
       shouldFreezeReactUpdates: () => pluginRegistry.store.options.freezeReactUpdates,
     });
 
-    const pendingShiftSelectionElement = createMemo((): Element | null => {
-      if (!isShiftMultiSelecting()) return null;
-      if (store.pendingCommentMode || isPendingContextMenuSelect()) return null;
-
-      const element = store.detectedElement;
-      if (!isElementConnected(element)) return null;
-      if (isRootElement(element)) return null;
-      if (store.frozenElements.includes(element)) return null;
-
-      return element;
+    const dragSelectors = createDragSelectors({
+      grab,
+      phase,
+      elementSelectors,
+      shiftMultiSelect,
+      isPendingContextMenuSelect,
+      debouncedDragPointer,
     });
-
-    const pendingShiftSelectionBounds = createMemo((): OverlayBounds | undefined => {
-      void viewportVersion();
-      const element = pendingShiftSelectionElement();
-      if (!element) return undefined;
-      return createElementBounds(element);
-    });
+    const {
+      isDraggingBeyondThreshold,
+      dragBounds,
+      dragPreviewBounds,
+      selectionBoundsMultiple,
+      frozenLabelEntries,
+      pendingShiftPreviewEntry,
+      cursorPosition,
+      shiftSelectionLabelMouseX,
+    } = dragSelectors;
 
     const toPageCoordinates = toPageCoordinatesUtil;
     const calculateDragDistance = (endX: number, endY: number) =>
       calculateDragDistanceUtil(store.dragStart, endX, endY);
     const calculateDragRectangle = (endX: number, endY: number) =>
       calculateDragRectangleUtil(store.dragStart, endX, endY);
-
-    const isDraggingBeyondThreshold = createMemo(() => {
-      if (!isDragging()) return false;
-      const dragDistance = calculateDragDistance(pointer().x, pointer().y);
-      return dragDistance.x > DRAG_THRESHOLD_PX || dragDistance.y > DRAG_THRESHOLD_PX;
-    });
 
     const isSpaceActivationKey = (event: KeyboardEvent) =>
       event.code === "Space" || event.key === " ";
@@ -376,119 +367,6 @@ export const init = (rawOptions?: Options): ReactGrabAPI => {
       actions.stopDragReposition();
       previousSpaceDragPointerPage = null;
     };
-
-    const dragBounds = createMemo((): OverlayBounds | undefined => {
-      void viewportVersion();
-
-      if (!isDraggingBeyondThreshold()) return undefined;
-
-      const drag = calculateDragRectangle(pointer().x, pointer().y);
-
-      return createFlatOverlayBounds(drag);
-    });
-
-    const dragPreviewBounds = createMemo((): OverlayBounds[] => {
-      void viewportVersion();
-
-      if (!isDraggingBeyondThreshold()) return [];
-
-      const pointer = debouncedDragPointer();
-      if (!pointer) return [];
-
-      const drag = calculateDragRectangle(pointer.x, pointer.y);
-      const elements = getElementsInDrag(drag, isValidGrabbableElement);
-      const previewElements =
-        elements.length > 0 ? elements : getElementsInDrag(drag, isValidGrabbableElement, false);
-
-      return previewElements.map((element) => createElementBounds(element));
-    });
-
-    const selectionBoundsMultiple = createMemo((): OverlayBounds[] => {
-      const previewBounds = dragPreviewBounds();
-      if (previewBounds.length > 0) {
-        return previewBounds;
-      }
-      const pendingBounds = pendingShiftSelectionBounds();
-      if (pendingBounds) {
-        return [...frozenElementsBounds(), pendingBounds];
-      }
-      return frozenElementsBounds();
-    });
-
-    const frozenLabelEntryAccessors = mapArray(
-      () => store.frozenElements,
-      (element) => {
-        const tagName = getTagName(element) || "element";
-        const componentName = getComponentDisplayName(element) ?? undefined;
-        return createMemo<FrozenLabelEntry | null>(() => {
-          void viewportVersion();
-          if (!isElementConnected(element)) return null;
-          const bounds = createElementBounds(element);
-          const anchorRatio = shiftMultiSelect.getAnchorRatio(element);
-          const mouseX =
-            anchorRatio === undefined ? undefined : bounds.x + bounds.width * anchorRatio;
-          return { tagName, componentName, bounds, mouseX };
-        });
-      },
-    );
-
-    const frozenLabelEntries = createMemo((): FrozenLabelEntry[] => {
-      if (isPromptMode() || store.frozenElements.length < 2) return [];
-      const entries: FrozenLabelEntry[] = [];
-      for (const readEntry of frozenLabelEntryAccessors()) {
-        const entry = readEntry();
-        if (entry !== null) entries.push(entry);
-      }
-      return entries;
-    });
-
-    const pendingShiftPreviewEntry = createMemo((): FrozenLabelEntry | null => {
-      if (isPromptMode()) return null;
-      const element = pendingShiftSelectionElement();
-      if (!element) return null;
-      void viewportVersion();
-      const tagName = getTagName(element) || "element";
-      const componentName = getComponentDisplayName(element) ?? undefined;
-      const bounds = createElementBounds(element);
-      return { tagName, componentName, bounds, mouseX: pointer().x };
-    });
-
-    const cursorPosition = createMemo(() => {
-      if (isCopying() || isPromptMode()) {
-        void viewportVersion();
-        const element = store.frozenElement || targetElement();
-        if (element) {
-          const center = getBoundsCenter(createElementBounds(element));
-          return {
-            x: center.x + store.copyOffsetFromCenterX,
-            y: store.copyStart.y,
-          };
-        }
-        return {
-          x: store.copyStart.x,
-          y: store.copyStart.y,
-        };
-      }
-      return {
-        x: pointer().x,
-        y: pointer().y,
-      };
-    });
-
-    const shiftSelectionLabelMouseX = createMemo((): number | undefined => {
-      if (!isShiftMultiSelecting()) return undefined;
-      if (store.frozenElements.length !== 1) return undefined;
-      void viewportVersion();
-
-      const element = store.frozenElements[0];
-      if (!isElementConnected(element)) return undefined;
-
-      const anchorRatio = shiftMultiSelect.getAnchorRatio(element);
-      if (anchorRatio === undefined) return undefined;
-
-      const bounds = createElementBounds(element);
-      return bounds.x + bounds.width * anchorRatio;
-    });
 
     createEffect(
       on(
